speak-bridge/server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pillow
numpy
elevenlabs==1.8.0